Observability: Refactoring to expose the MDS profile
