+++ observability/src/main/java/no/nordicsemi/android/observability/log/Category.kt
@@ -1,0 +1,10 @@
+package no.nordicsemi.android.observability.log
+
+import no.nordicsemi.kotlin.log.Log
+
+/**
+ * The Log category emitted by this library.
+ */
+enum class Category: Log.Category {
+    MDS,
+}
